Changed so that TowerDefence is the main class
