--- src/main/java/com/group34/Model/Game/Repaintable.java
@@ -1,6 +1,0 @@
-package com.group34.Model.Game;
-
-public interface Repaintable {
-    static final int FPS = 60;
-    void repaint();
-}
